refactor: rename `background` module to `task` in `tcads-client`
